requirements.txt
# Core PyTorch dependencies
torch>=2.0.0
torchvision>=0.15.0

# Configuration and utilities
pyyaml>=6.0

# Image processing
opencv-python>=4.7.0.72
numpy>=1.24.0
Pillow>=9.5.0

# Metrics and evaluation
scikit-image>=0.20.0

# Optional: For better progress tracking during training
tqdm>=4.65.0